refactor: xk6 builder API removed
